RLIB-16: Better fix for the previous one.
When UTF-8 is enabled, r_strlen() is g_utf8_strlen() that returns
a signed value. When UTF-8 is disabled, r_strlen() is plain strlen()
and it returns an unsigned value. Cast the left side of the comparison
to signed and modify the type of "n" to ssize_t to keep it signed
and make it change its size just like size_t and glong.