RLIB-16: RLIB-16: Fix memory leaks shown by src/examples/c/relprt.c
The report XML parsed by libxml2 was not freed which was the biggest
remaining leak. Another small leak was also fixed in parsexml.c.

Now the parser tries to be more resilient against malformed XMLs.
Instead of ignoring unknown XML elements, it rejects them and
passes the error back to the caller. The locally allocated data
is freed in error paths. This should be tested further with more
unit tests.

Currently there is no leak in src/examples/c/relprt.c when the
queries are omitted from the report, i.e.: ./relprt -n