RLIB-26: Add a new member function to struct input_filter
This is to return the number of rows in the resultset, which is
needed to properly track followers later.

Remove the "forward only" knowledge from the ODBC input driver
and always copy the rows / fields and count the rows. This is
because some ancient ODBC drivers don't report the number of rows
in te result set and if the result set is implemented as a cursor
in the driver, we may not know the total number of rows without
traversing the result set.