RLIB-26: Fixed recordset row traversal when 1:1 or n:1 followers are present
Datasources were re-written (again) so the first step is equivalent
to SQL cursors, i.e. they really start BEFORE the first row, so
empty datasets can be detected properly. This was an off-by-one
bug before.

N:1 followers really produce multiple rows with multiple matches
against the leader query. When no matches are present, the right
side is empty as in an SQL LEFT OUTER JOIN.

New test cases were added to exercise both 1:1 and n:1 followers
in multiple ways.

Breaks are known to be broken now as they want to go back in the
resultset which was not performed properly before with followers
anyway.

TODOs:
1. Cache the last row and modify rlib_execute_pcode() so it can work
   from the cached data, thereby removing the necessity to go
   backwards in resultsets.
2. Another input datasource surgery. Only restarts and forward
   traversal will be present, which will allow using a cursor in the
   PostgreSQL datasource. It will be a huge space saving feature
   balanced with somewhat lower performance.
3. The ODBC input source will restore using SQLFetchScroll straight
   instead of caching rows from the resultset.