RLIB-26: Breaks seem to work now
Double caching (one in struct rlib_break_fields, via the rval2 and
rval members) is now eliminated, the row traversal and row caching
in the resultsets are enough to produce correct values for breaks.

The rlib_value caching in struct rlib_break_fields was suspicious
anyway, it checked that the value is of the string type. Now this
is completely generic, as break fields can be arbitrary expressions.

Refreshed the Sicom INC address in configure.ac

Fixed memory leaks in the TXT report output driver and ones
I introduced earlier with the row caching code.

Some manual reindentation / code tidying was also done as I was
reading the code.