RLIB-26: Allow mixing different (delayed) variables in the same expression
Introduced OPERAND_VALUE to replace OPERAND_VARIABLE in a pcode
with the variable's final (or: before reset) value so variables
that are reset on different breaks can occur in the same expression.

Fixed premature substitution of delayed expressions in output drivers
to allow multiple delayed variables work in the same expression.

src/examples/c/breaks2.c and src/examples/php/breaks2.xml exercise this
new feature.

Fixed a previously introduced memory leak in the HTML output driver.

There is a leak now when running the "graph" example, fixing is TODO.