RLIB-26: Fix printing more report variables
Variables without either resetonbreak="some_break" or precalculate="yes"
XML parameters were not printed correctly. They were either missing
or used the last value at the end of the report.

Also renamed the "immediate" XML parameter back to "precalculate"
and the internal handling is negated. The reason is that there are
a lot of reports out there that expect the old naming. The behaviour
of an "immediate" variable is exactly the opposite of a "precalculated"
one, so it was simple.

src/examples/php/breaks2.xml was modified to reflect this change.