RLIB-21: Port RPDF to libharu
libharu is an open source PDF generator that allows font embedding
out of the box. RPDF is now a thin wrapper over libharu to simulate
the previous RPDF behaviour and quirks. Switching to libharu gives
better PNG image support in RLIB for free, too. The visual look of
the PDF is identical to the old PDF generated by RPDF, the file
size is smaller as libharu generates more optimized PDF layout.

PDF generation time is proportional to the number of embedded fonts.