RLIB-21: Make creation of delayed data list faster
Use g_slist_prepend instead of g_slist_append when constructing
the delayed data list for breaks and for the end of report.
The order of printing delayed data is unimportant but the speed
of the library is. Long reports may get slower if delayed data
is used for detail rows.