RLIB-23: Fix the test case generator for multiple runs
In case the same process runs reports multiple times, older
generated test cases would be overwritten. This is a problem when
PHP-FPM is used and the reports use the PHP module.

The test case name is now "testcase-PID-COUNT.c", the PID is the
process ID and COUNT is an ever increasing counter. This counter
is protected by a mutex to make it MT-safe.