RLIB-23: Fix possible crash when opening datasources
If the database is not available (MySQL, PostgreSQL, ODBC) and
the datasource modules were built as external modules, opening
the datasource failed. At this point the datasource module was
to be unloaded (or its refcount decreased by 1) by g_module_unload()
and the input->free() function was to be called. Since the module
was already unloaded, the input pointer was left lingering and
calling into the unloaded module caused the crash. Calling ->free()
earlier than unloading the module fixes this particular crash.