RLIB-16: Fix missing pie charts in PDF when built with -O2 -D_FORTIFY_SOURCE=2
rpdf_make_page_stream() has a wrong usage pattern for sprintf():
the target buffer (statically allocated on the stack!) was also
used as a source string, making it unreliable, *printf() in GLIBC
with optimized flags becomes a macro.

Use a GString and g_string_append_printf() to perform the intended
operation in a reliable way. This way realloc is used implicitly.