RLIB-17 / RLIB-24: Allow relative files
RLIB now prefers files relative to the path of the report XML
specified with rlib_add_report() instead of the current working
directory.

The input datasource internal API was modified so the "rlib *r"
pointer was added to struct input_filter and every
    xxx_new_input_filter(void);
function was changed to
    xxx_new_input_filter(rlib *r);

This was needed to use the relative file path discovery from
the XML and CSV data sources.

A new API call was also added:

gint rlib_add_search_path(rlib *r, const gchar *path);

This is available from C and PHP for now.

Report XML parts, images, CSV and XML data source files are
searched in this order, the first one is returned:
1. relative to the main report XML
2. extra search paths, if there are more than one of them,
   they are searched in the order they were added
3. relative to the process current working directory,
   this is for keeping compatibility with older versions

To make unit testing easier, the "No queries added to report"
condition was demoted to warning from fatal condition. This allows
running reports with only a report header or page header that
exercises certain behaviour elements, like images or alignment.

The previously separate in-house issue called RLIB-24, a bug fix
for HTML output over RLIB-17 was merged in this commit, as well as
another crash caused by the relaxed check whether there are queries
added to the report.