Converted the build framework of the RLIB php module to phpize
rlib.h used <libxml/parser.h> which is used by the PHP module
which doesn't need to be linked to LIBXML directly. The result
was a build failure in the PHP module.

The fix is to sanitize rlib.h a little:
* Don't #include <libxml/parser.h>
* "xmlChar *xml" from struct rlib_from_xml was unrolled as
  "unsigned char *xml"
* "xmlDocPtr doc" in struct rlib_report became gpointer

C sources were adjusted because <libxml/parser.h> implicitly
includes <stdio.h> and <stdlib.h>.

libsrc/axis.c was reindented a little because of this warning
when built with GCC 6.x (Fedora 24, Yocto 2.2) that causes
a build error with -Werror:

axis.c: In function 'adjust_limits':
axis.c:430:1: error: this 'if' clause does not guard... [-Werror=misleading-indentation]
  if (sd<1 || sd >4)
  ^~
axis.c:434:2: note: ...this statement, but the latter is misleadingly indented as if it is guarded by the 'if'
   if(denyMinEqualsAdjMin && dataMin < 0 && dataMax < 0)
   ^~