Reference files from the source directories in custom make targets
When cross-compiling under Bitbake, the build has the following
scenario:

1. extract the source tarball
2. create a build directory in the same toplevel directory
3. enter the build directory and execute the configure script
   in this way:

   ../rlib-2.0.x/configure <options>

This scenario failed in src/examples/c. The Makefile target was:

products.xml:
	$(LN_S) ../common/products.xml

The error message was:

ln -s ../common/products.xml
ln: failed to create symbolic link './products.xml': File exists

The reason is subtle. During "make", the symlink didn't exist and
it was created. During "make install", although the symlink existed,
it pointed to a non-existing file because the source and build
directories are not the same. The make command looks for actual
files, i.e. the targets behind the symlink, so it executed the
above Makefile target again for products.xml and the "ln -s" command
failed because the symlink already existed.

The fix is to use $(srcdir) or $(abs_srcdir) to reference such
files in the custom Makefile targets. This includes the subdirectories
in "src/examples" and the "doc" directory in the RLIB source tree.