Handle RLIB_DEBUGGING and RPDF_DEBUGGING environment variables
When RPDF_DEBUGGING is set, RLIB will use the UNIX epoch for
CreationDate in the PDF output, so the result can be compared
for regression testing.

When RLIB_DEBUGGING is set, the HTML output will use a simple
counter in the name of the generated images, so the result can be
compared for regression testing. The counter variable was moved
from gd.c::unique_file_name() (where it was a local static variable)
to the HTML output code in the private structure. This reduces
the amount of threading bugs in RLIB.