Reference RLIB build directory in bindings/php/config.m4
The previous commit, although descended properly to the RLIB
source directory to build the RLIB PHP module, it was not enough.

config.m4 referenced ../../rpdf/.libs and ../../libsrc/.libs
to link to recently built RPDF and RLIB libraries but when
the source and build directories are not the same, linking
the PHP module fails.

Allow passing the toplevel build directory to the PHP module
via a new config.m4 option: --with-rlib-builddir=...