Fix the previous commit
The changes to rlib_alloc_query_space() needed a follow up change
in rlib_add_query_pointer_as() to prevent a double accounting of
the number of queries in the report, using the return value
of rlib_alloc_query_space() for a slight optimization. Without this,
the previous change caused crashed when running the queries.