RLIB-22: Fix crash in PHP 7 support
The way variables are stored in the hash of variables and
referenced have changed significantly in PHP 7.x and refcounted
zvals have been introduced a while ago.

This (Z_TYPE_P() == IS_REFERENCE) was not handled before.
This resulted in returning NULL from php_array_new_result_from_query()
when rlib_execute_queries() was executed the second time.

This in turn caused a crash in resolution.c.

I took this opportunity to sanitize php_array_new_result_from_query()
a little and bring it in line with the RLIB 2.0.15 / 3.0.0 code.