RLIB-23: Create test cases automatically
When RLIB_CREATE_TEST_CASE=1/on/true is set in the environment
(either in the UNIX shell or as a PHP variable) then data, the
report XML(s) and the API calls are recorded and dumped into
a standalone C test case plus a test case metadata file.
Since the dumped test case is in C, it exploits the previously
added C array datasource.

This way it's ready to be used with the regression test framework.
It is good for create a huge set of regression tests from the
existing set of reports easily.

For this to work, two infrastructure inside RLIB were extended:

1. The input datasources gained two more API functions:
   num_fields() and get_field_name().

2. The environment framework (shell, Python and PHP variables)
   can dump all non-array variables into a memory string.

The ODBC datasource now uses the same coding for returning field
indexes as all other datasources.