RLIB-21: Fix non-convertible characters in the output
Instead of using g_convert_with_iconv(), use g_iconv() and handle
the fallouts, e.g. pre-allocate the space for the converted string
like g_convert_with_iconv() does internally.

We don't want to see !ERR_ENC2 printed in the output. If g_iconv()
returns -1, skip the character in the input and replace it with
a space in the output and continue with the rest.