RLIB-29: Allow using doubled single quotes in strings
The standard escaping/quoting of the same quote character inside
a string that is also used to indicate the string is to double it.

Examples:
  str = 'something''s wrong'
  str = "Joe said ""I will go."""

Since RLIB uses single quotes to indicate a string constant,
allow using a doubled single quote inside it.

It is to allow using e.g.: format="'!#%''.2d'", i.e. with the
new style format specifier, include the thousand separators
explicitly when printing a number with two decimal places.

This is a very loose support and only a workaround for a missing
piece. A single quote is equally supported inside a string which
is obviously a bug but it's hard to fix in the current parser.
The real fix would be to port the whole expression handling to
a Flex and Bison based parser.